requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
idna==3.10
keyboard==0.13.5
numpy==2.1.2
packaging==24.1
pefile==2023.2.7
pillow==11.0.0
plyer==2.1.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pywin32-ctypes==0.2.3
requests==2.32.3
urllib3==2.2.3
